Compute the inverse of matrices using iterative algorithm
There are three main part of the iterative algorithm:
	-- Initial value. In the experiments, we use two kinds of initial values.
	The first one is from the analog ciruitry, which are obtained by truncating
	the precision of the inverse of the target matrix; The second one is computed
	by digital computer. Two methods are used here: Pan and Ben.
	-- Update function. Xn+1 = Xn*(2I-A*Xn).
	-- Stopping criterion. The experiment draws figures of the norm of the error
	matrix over iterations and then compute the error threshold which indicates
	convergence of the iterative algorithm. For different sizes, the stopping
	criterion is also different.
MatrixInverseIterative_rand.m perform number of experiment on a single size of matrices.
dataProcess.m processes the data of one size and draws some figures.
run_exp.m executes the experiments for different sizes of matrices, which should be the
main program to run.
Note that the experiment result (figures and some data) is not included in the repo. The main
reason is that the amount of figures are so big that it seems to take forever to add them.